Refactors project structure and dependencies (#1)
* Refactors user repository and achievement flow

Changes the UserRepository interface and its implementation to observe achievements progress as a Map instead of a List.

This allows for more efficient access and management of achievement data.

* update readme

* refactor

* refactor usecases

* refactor user usecases

* Refactors data layer and adds user repository

This commit refactors the data layer by removing DTOs and mappers, and moves repositories to the data package.

It also introduces a UserRepository to manage user-related data,
including profile, settings, stats, inventory, achievements and task history.

It implements the creation of a new user.

It simplifies data access by directly using domain models with Firebase.

* Refactors achievement model and logic.

Moves achievement-related logic into dedicated classes and introduces an event-driven system.
This improves modularity and enables a more flexible achievement evaluation process.
Specifically:
- Renames Achievement to AchievementDisplay, adding achieved state
- Introduces AchievementEvent and AchievementRule to define conditions for unlocking achievements
- Creates AchievementEngine for evaluating rules and triggering achievements
- Introduces GetAchievementsUseCase to retrieve and display achievements
- Removes unused AchievementRequirement

* Refactors dependency injection modules.

Streamlines the Koin modules by removing unused module dependencies and adjusting constructor parameters for view models and use cases. This simplifies the dependency graph and improves code maintainability.

* Refactors auth and user management logic

Changes auth repository to return user ID upon registration.

Updates user related use cases to correctly handle scenarios where the user is not logged in by returning an Error instead of a ValidationError.

Refactors all use cases to return unit.

Fixes return of wrong xp calculation

* Removes unused Firestore extension function

Removes the DocumentReference.observeList() extension function, as it appears to be unused and potentially incorrect in its implementation. This simplifies the Firestore utility extensions and reduces the risk of future errors.
Also corrects a typo in a comment in RootView.

* Refactors dependency injection and UI elements

Refactors the dependency injection setup by introducing a dedicated UserModule for user-related dependencies, promoting better organization and modularity.

Updates UI elements, specifically the login, registration, header and custom select components, by adjusting corner shapes for a more consistent and modern look and feel.

Changes the color scheme to improve visual appeal.

* Updates project documentation and structure

* Adds structure to the project

Adds comments for clarity
Adds constants file and extension methods
Refactors project modules

+++ app/src/main/java/marcel/uni/gamifiedplanner/domain/shop/model/ShopItem.kt
@@ -4,6 +4,6 @@ data class ShopItem(
     val id: String = "",
     val name: String = "",
     val description: String = "",
-    val price: Int = 0,
+    val price: Long = 0,
     val iconResId: Int? = null,
 )
